requirements.txt
crewai>=0.12.2
streamlit>=1.31.1
PyPDF2>=3.0.1
requests>=2.31.0
python-dotenv>=1.0.0
pypdf
PyMuPDF
langchain
langchain-openai
pytesseract
Pillow
numpy 